Modifications on power

+++ 67_pow.c
@@ -1,6 +1,6 @@
 // Make your own pow function.
 #include <stdio.h>
-#include<math.h>
+#include <math.h>
 int main()
 {
     int power(int x, int y);
@@ -13,9 +13,10 @@ int main()
 
 int power(int a, int b)
 {
-    int result=1;
-    for(int i=1;i<=b;i++){
-        result = result *a;
+    long long int result = 1;
+    for (int i = 1; i <= b; i++)
+    {
+        result = result * a;
     }
     return result;
 }
